+++ NEWS.md
@@ -5,6 +5,7 @@ Updates:
   * Corrects regularise_time output for "day" and "hour"
   * Makes sure reg_SL_tbl class is returned by regularise_time
   * Add blocks to default title of vis_NAs plot
+  * SL_tbl_load now guesses schema from table
   
 # studentlife 1.0.3
 This release only updates the authorship to give